# dist/api-docs.json
{
  "swagger": "2.0",
  "basePath" : "/api",
  "tags" : [
    {
    "name": "Node",
    "description": "Basic Rest"
    }
  ],
  "info": {
  "version": "0.0.1",
  "title": "ZedWave"
  },
  "paths": {
    "/nodes": {
      "get": {
        "tags": ["Node"],
        "description": "Gets `Node` objects.",
        "parameters": [
            {

            }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "title": "ArrayOfNodes",
              "type": "array",
              "items": {
                "title": "Node",
                "type": "object",
                "properties": {
                  "__v": {
                    "type": "integer"
                  },
                  "_id": {
                      "type": "string"
                  },
                  "classes": {
                      "type": "array"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/nodes/{nodeId}": {
      "get": {
        "tags": ["Node"],
        "description": "Gets single `Node` object.",
        "parameters": [
          {
            "in": "path",
            "name": "nodeId",
            "description": "ID of node that needs to be fetched",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "title": "ArrayOfNodes",
              "type": "array",
              "items": {
                "title": "Node",
                "type": "object",
                "properties": {
                  "__v": {
                    "type": "integer"
                  },
                  "_id": {
                      "type": "string"
                  },
                  "classes": {
                      "type": "array"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Node"],
        "description": "Deletes `Node` object.",
        "parameters": [
          {
            "in": "path",
            "name": "nodeId",
            "description": "ID of node that needs to be fetched",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
              "description": "Successful response"
          }
        }
      }
    },
    "/nodes/{nodeId}/classes": {
      "get": {
        "tags": ["Node"],
        "description": "Gets single `Node` object's classes.",
        "parameters": [
          {
            "in": "path",
            "name": "nodeId",
            "description": "ID of node that needs to be fetched",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
              "description": "Successful response"
          }
        }
      }
    }
  }
}